PR improvements
- Move new mage target to elastic-agent-libs.
- Create a single binary at the root and make loadgenerator a
  subcommand.
