--- cmd/loadgenerator/load-generator.yml
@@ -1,19 +1,0 @@
-# Delay between messages are sent, default is 1 second
-delay: "0.5s"
-
-# Time delta between messages (affects @timestamp field) but not
-# the throughput. Default is 1s
-timedelta: "0.5s"
-
-# Execute indefinitely, if set, 'eventscount' is ignored.
-# Default is true
-loop: true
-
-# Number of events to generate
-eventscount: 100
-
-logging:
-  level: debug
-  selectors: ["*"]
-  to_stderr: true
-  to_files: false
