+++ pkg/config/config.go
@@ -5,28 +5,84 @@
 package config
 
 import (
-	"github.com/elastic/elastic-agent-libs/logp"
+	"flag"
+	"fmt"
+	"os"
+	"path/filepath"
 
-	// This adds the global log flags
-	_ "github.com/elastic/elastic-agent-libs/logp/configure"
+	"github.com/elastic/elastic-agent-inputs/inputs/loadgenerator"
+	"github.com/elastic/elastic-agent-libs/config"
+	"github.com/elastic/elastic-agent-libs/logp"
 )
 
-// TODO (Tiago): make those global, at the moment they're defined
-// in the load-generator input
-// var (
-// 	ConfigPath     string
-// 	ConfigFilePath string
-// )
+const (
+	defaultConfigName = "agent-inputs.yml"
+)
 
-// // Define some flags used by Elastic-Agent and its friends
-// func init() {
-// 	fs := flag.CommandLine
-// 	fs.StringVar(&ConfigFilePath, "c", defaultConfigName, "Configuration file, relative to path.config")
-// 	fs.StringVar(&ConfigPath, "path.config", ConfigPath, "Config path is the directory Inputs looks for its config file")
-// }
+var (
+	ConfigPath     string
+	ConfigFilePath string
+)
+
+// Define some flags used by Elastic-Agent and its friends
+func init() {
+	fs := flag.CommandLine
+	fs.StringVar(&ConfigFilePath, "c", defaultConfigName, "Configuration file, relative to path.config")
+	fs.StringVar(&ConfigPath, "path.config", ConfigPath, "Config path is the directory agent-inputs looks for its config file")
+}
 
 // Config defines the options present in the config file
-// TODO (Tiago): fix config parsing/unpacking
 type Config struct {
-	Log logp.Config `yaml:"logging" json:"logging"`
+	Log           logp.Config          `yaml:"logging" json:"logging"`
+	LoadGenerator loadgenerator.Config `yaml:"loadgenerator" json:"loadgenerator"`
+}
+
+// ReadConfig returns the populated config from the specified path
+func ReadConfig() (Config, error) {
+	path := Filepath()
+
+	configYAML, err := os.ReadFile(path)
+	if err != nil {
+		return Config{}, fmt.Errorf("error reading input file '%s': %w", path, err)
+	}
+
+	raw, err := config.NewConfigWithYAML(configYAML, "")
+	if err != nil {
+		return Config{}, fmt.Errorf("error parsing yaml config: %w", err)
+	}
+
+	config := defaultConfig()
+	if err := raw.Unpack(&config); err != nil {
+		return config, fmt.Errorf("error unpacking shipper config: %w", err)
+	}
+
+	return config, nil
+}
+
+// Filepath returns the default config file path or the config
+// file path set as flags
+func Filepath() string {
+	if ConfigFilePath == "" || ConfigFilePath == defaultConfigName {
+		return filepath.Join(ConfigPath, defaultConfigName)
+	}
+	if filepath.IsAbs(ConfigFilePath) {
+		return ConfigFilePath
+	}
+	return filepath.Join(ConfigPath, ConfigFilePath)
+}
+
+func defaultConfig() Config {
+	logCfg := logp.DefaultConfig(logp.DefaultEnvironment)
+	logCfg.Beat = "agent-inputs"
+	logCfg.ToStderr = true
+	logCfg.ToFiles = false
+
+	// While we're in development, keep the default debug
+	logCfg.Level = logp.DebugLevel
+	logCfg.Selectors = []string{"*"}
+
+	return Config{
+		Log:           logCfg,
+		LoadGenerator: loadgenerator.DefaultConfig(),
+	}
 }
